fix: StarRating files naming, RatingForm design
