+++ src/styles/RatingForm.scss
@@ -5,7 +5,7 @@
         align-items: center;
         justify-content: center;
         width: 100%;
-        margin: 100px 0;
+        margin: 100px 0 50px;
     }
 
     &__group {
@@ -60,6 +60,7 @@
 
     textarea {
         margin: 15px;
+        height: 200px;
     }
 
 }
